flask_API/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
flask
flask_socketio
quart
gunicorn
waitress
eventlet
gevent
opencv-python
numpy
av
spacy
nltk
scikit-learn
PyPDF2
matplotlib
language_tool_python
torch
transformers
fasttext-wheel
python-dotenv
requests
moviepy
pywin32
mediapipe
ffmpeg-python
pydub
torchaudio
IPython
scikit-image
librosa
SpeechRecognition
websocket-client